Getting encoding for mzML only when needed
Getting file metadata, BPC/TIC does not require encoding. Also getting chromatograms (to be suggested) can require different encoding. An error is given if no spectra are present, such as the case of mzML files with only chromatograms (SRM, MRM, or SIM). This would prevents the error but a better solution could be implemented.